+++ docs/history.md
@@ -42,6 +42,8 @@ This is the first stable release of Hatch v1, a complete rewrite. Enjoy!
 
 ### Unreleased
 
+### [0.25.0](https://github.com/pypa/hatch/releases/tag/hatchling-v0.25.0) - 2022-05-15 ### {: #hatchling-v0.25.0 }
+
 ***Added:***
 
 - Add `skip-excluded-dirs` build option